модуль category  test_category
