+++ htmlcov/class_index.html
@@ -55,7 +55,7 @@
         </h2>
         <p class="text">
             <a class="nav" href="https://coverage.readthedocs.io/en/7.9.2">coverage.py v7.9.2</a>,
-            created at 2025-07-09 12:36 +0300
+            created at 2025-07-09 12:41 +0300
         </p>
     </div>
 </header>
@@ -113,40 +113,48 @@
                 <td class="right" data-ratio="9 9">100%</td>
             </tr>
             <tr class="region">
-                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html#t9">src\main.py</a></td>
-                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html#t9"><data value='Category'>Category</data></a></td>
+                <td class="name left"><a href="z_145eef247bfb46b6_category_py.html#t4">src\category.py</a></td>
+                <td class="name left"><a href="z_145eef247bfb46b6_category_py.html#t4"><data value='Category'>Category</data></a></td>
                 <td>17</td>
                 <td>0</td>
                 <td>0</td>
                 <td class="right" data-ratio="17 17">100%</td>
             </tr>
             <tr class="region">
-                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html#t52">src\main.py</a></td>
-                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html#t52"><data value='CategoryIterator'>CategoryIterator</data></a></td>
+                <td class="name left"><a href="z_145eef247bfb46b6_category_py.html">src\category.py</a></td>
+                <td class="name left"><a href="z_145eef247bfb46b6_category_py.html"><data value=''><span class='no-noun'>(no class)</span></data></a></td>
+                <td>13</td>
+                <td>0</td>
+                <td>0</td>
+                <td class="right" data-ratio="13 13">100%</td>
+            </tr>
+            <tr class="region">
+                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html#t12">src\main.py</a></td>
+                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html#t12"><data value='CategoryIterator'>CategoryIterator</data></a></td>
                 <td>10</td>
                 <td>0</td>
                 <td>0</td>
                 <td class="right" data-ratio="10 10">100%</td>
             </tr>
             <tr class="region">
-                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html#t79">src\main.py</a></td>
-                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html#t79"><data value='Smartphone'>Smartphone</data></a></td>
+                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html#t39">src\main.py</a></td>
+                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html#t39"><data value='Smartphone'>Smartphone</data></a></td>
                 <td>5</td>
                 <td>0</td>
                 <td>0</td>
                 <td class="right" data-ratio="5 5">100%</td>
             </tr>
             <tr class="region">
-                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html#t107">src\main.py</a></td>
-                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html#t107"><data value='LawnGrass'>LawnGrass</data></a></td>
+                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html#t67">src\main.py</a></td>
+                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html#t67"><data value='LawnGrass'>LawnGrass</data></a></td>
                 <td>4</td>
                 <td>0</td>
                 <td>0</td>
                 <td class="right" data-ratio="4 4">100%</td>
             </tr>
             <tr class="region">
-                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html#t131">src\main.py</a></td>
-                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html#t131"><data value='Order'>Order</data></a></td>
+                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html#t91">src\main.py</a></td>
+                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html#t91"><data value='Order'>Order</data></a></td>
                 <td>4</td>
                 <td>0</td>
                 <td>0</td>
@@ -155,10 +163,10 @@
             <tr class="region">
                 <td class="name left"><a href="z_145eef247bfb46b6_main_py.html">src\main.py</a></td>
                 <td class="name left"><a href="z_145eef247bfb46b6_main_py.html"><data value=''><span class='no-noun'>(no class)</span></data></a></td>
-                <td>65</td>
+                <td>55</td>
                 <td>29</td>
                 <td>0</td>
-                <td class="right" data-ratio="36 65">55%</td>
+                <td class="right" data-ratio="26 55">47%</td>
             </tr>
             <tr class="region">
                 <td class="name left"><a href="z_145eef247bfb46b6_print_mixin_py.html#t4">src\print_mixin.py</a></td>
@@ -205,10 +213,10 @@
             <tr class="total">
                 <td class="name left">Total</td>
                 <td class="name left">&nbsp;</td>
-                <td>192</td>
+                <td>195</td>
                 <td>41</td>
                 <td>0</td>
-                <td class="right" data-ratio="151 192">79%</td>
+                <td class="right" data-ratio="154 195">79%</td>
             </tr>
         </tfoot>
     </table>
@@ -220,7 +228,7 @@
     <div class="content">
         <p>
             <a class="nav" href="https://coverage.readthedocs.io/en/7.9.2">coverage.py v7.9.2</a>,
-            created at 2025-07-09 12:36 +0300
+            created at 2025-07-09 12:41 +0300
         </p>
     </div>
     <aside class="hidden">
