+++ htmlcov/index.html
@@ -54,7 +54,7 @@
         </h2>
         <p class="text">
             <a class="nav" href="https://coverage.readthedocs.io/en/7.9.2">coverage.py v7.9.2</a>,
-            created at 2025-07-09 12:36 +0300
+            created at 2025-07-09 12:41 +0300
         </p>
     </div>
 </header>
@@ -92,11 +92,18 @@
                 <td class="right" data-ratio="9 12">75%</td>
             </tr>
             <tr class="region">
+                <td class="name left"><a href="z_145eef247bfb46b6_category_py.html">src\category.py</a></td>
+                <td>30</td>
+                <td>0</td>
+                <td>0</td>
+                <td class="right" data-ratio="30 30">100%</td>
+            </tr>
+            <tr class="region">
                 <td class="name left"><a href="z_145eef247bfb46b6_main_py.html">src\main.py</a></td>
-                <td>105</td>
+                <td>78</td>
                 <td>29</td>
                 <td>0</td>
-                <td class="right" data-ratio="76 105">72%</td>
+                <td class="right" data-ratio="49 78">63%</td>
             </tr>
             <tr class="region">
                 <td class="name left"><a href="z_145eef247bfb46b6_print_mixin_py.html">src\print_mixin.py</a></td>
@@ -123,10 +130,10 @@
         <tfoot>
             <tr class="total">
                 <td class="name left">Total</td>
-                <td>192</td>
+                <td>195</td>
                 <td>41</td>
                 <td>0</td>
-                <td class="right" data-ratio="151 192">79%</td>
+                <td class="right" data-ratio="154 195">79%</td>
             </tr>
         </tfoot>
     </table>
@@ -138,7 +145,7 @@
     <div class="content">
         <p>
             <a class="nav" href="https://coverage.readthedocs.io/en/7.9.2">coverage.py v7.9.2</a>,
-            created at 2025-07-09 12:36 +0300
+            created at 2025-07-09 12:41 +0300
         </p>
     </div>
     <aside class="hidden">
